backend/requirements.txt
beanie
black
fastapi[standard]
pydantic-extra-types
pydantic-settings
pyjwt
passlib[bcrypt]